(#62) Remove links to Nuget Gallery
Within the source code for Chocolatey Gallery, there are a number of
links out to Nuget Gallery (as a result of using the Nuget Gallery code
as a basis for Chocolatey Gallery).  Where possible, these have been
removed.  In addition, the footer of the 404 and 500 error pages have
been updated to be consistent with the rest of the site.