(GH-17) Save package files
When a package is pushed during inspection, package files and content should
be saved.

Files that are binary will be saved, but the content of the file will not.
The files are determined to be text files based on a white-listed
group versus trying to determine if the file is text at runtime.