(GH-88) Deprecate using custom route handler
Originally we needed a custom route handler so that urls behind the gateway
were built appropriately, but there is new evidence that simply setting
aspnet:UseHostHeaderForRequestUrl in the appSettings will alleviate this issue.

The config has already contained this value for a little while, now switching
over to exclusively using it.