(GH-98) Save an existing package
This allows an already package to be saved with all of the proper information
being updated. This is only good for scenarios where a package is in moderation
and the user needs to resubmit the package. They would continue to upload the
same one with the same version until it is approved or rejected.

The existing code really needed worked over.