(GH-135) Do not convert description to code based on indentation
Use a custom MarkdownSharp generator that includes the option `SkipCodeBlocks`.

There is a rather difficult to decipher bug in MarkdownSharp regarding the use
of MarkdownOptions. The short of it is, **DO NOT USE MarkdownOptions**. The
behavior is incorrect and will not produce good html. Prefer instead to use
object initializer or set the options separately after instantiation. Here's
the bug for details including a unit test -
https://code.google.com/p/markdownsharp/issues/detail?id=67