(GH-148) Don't show awaiting moderation msg if latest
If a user (not a moderation role) is on the latest version available
and the packages awaiting moderation are older versions, do not show
the message `There are versions of this package awaiting moderation`.
It gives the appearance to users that there may be newer versions in
the pipeline, when in fact there are not.