(GH-195) UrlExtensions.PackageDownload Fixes
PackageDownload was using http as the scheme, which would later get
redirected to an https request by the site.  It was also removing the
port, which is an older hack we needed for AppHarbor before we moved to
the simpler appSettings key `aspnet:UseHostHeaderForRequestUrl`. This
was done in 98f760e94012ebae96bbe7272de01f04ed8d74fc, but this was
never corrected at that time.