(build) ensure the call to build.bat works
When VS2012/2013 is installed, it makes the call to build.bat stop
working with the following error:

~~~
C:\codelocal\chocolatey.org\chocolatey\Website\Website.csproj(1513,3):
error MSB4019: The imported project "C:\Program Files
(x86)\MSBuild\Microsoft\VisualStudio\v11.0\WebApplications\Microsoft.
WebApplication.targets" was not found. Confirm that the path in the
<Import> declaration is correct, and that the file exists on disk.
~~~

This is due to a movement of the WebApplications targets file under v12
folder instead of v11.

In the project file itself you can see it is preferring the new
location:

~~~xml
<PropertyGroup>
    <VisualStudioVersion Condition="'$(VisualStudioVersion)' ==
''">10.0</VisualStudioVersion>
    <VSToolsPath Condition="'$(VSToolsPath)' ==
''">$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)</VSToolsPath>
  </PropertyGroup>
  <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
  <Import
Project="$(VSToolsPath)\WebApplications\Microsoft.WebApplication.targets"
Condition="'$(VSToolsPath)' != ''" />
  <Import
Project="$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v10.0\WebApplications\Microsoft.WebApplication.targets"
Condition="false" />
~~~

However removing the last line there as recommended in some of the
researched answers makes VS2010 think it is time to upgrade the project
and some answers state it does more harm than good.

The fix is to either add `/p:VisualStudioVersion=12.0` so that it will
prefer the specified location during command line build or point it to
using the Visual Studio built-in MSBuild if available with a fall back
to the old .NET Framework version of MSBuild if it is not found.