(GH-211) Hide install/upgrade badge unless listed
An unlisted package can still be installed if a user passes the name
AND version of the package. But it will not show up in search results
and will not be installable without also providing the version.

If a package is not listed, there is a reason for that, from being
unlisted by the maintainer/site admins/moderators, to not being
approved or possibly rejected for a multitude of reasons. An
unsuspecting user that is able to go directly to the package page for a
package through possibly the result of a google search or some other
way may see the install/upgrade section and not understand that the
package has been unlisted or may miss the reasons and attempt to
install the package with varying degrees of success.

Without the install/upgrade section, there is no command the user can
pick up to run. And a regular user won't have access to know what
versions were there, only the listed versions. This makes it much
harder for a user to decide to move forward to install a package,
because it would be impossible for them without guessing at the
versions and getting it correct. That level of effort is not something
that normal users may go to the trouble of versus simply copying and
pasting a command to run.