(GH-219) Do not send email if same
If a user updates their email address to the same email, or just
updates their account information, sometimes it will attempt to send an
email change notice to the same email address. This is an annoying
issue.

If upon checking, you find the email addresses are the same, don't send
the notice.