(GH-226) Always show published date as last updated
It's annoying to see the last updated date where the published date
should be, last updated is affected by many things including lucene
indexing so it's not really something that should be shown to users.
Instead use Published date, which gives the date a package was pushed
up by a maintainer.