(GH-253) Update package test result status
When test results come in, the package test status should be updated
for display on the page, including any updates on the result details
url. This will drive the view on the pages for test results.

If the package is in a submitted status, update the package review
comments. If the package test has failed, change the submitted status
to waiting and ensure the maintainer knows that they will need to take
action to continue the review process.

If the package is not in a submitted status and the package has failed,
send the maintainer an email about the failure.