(GH-269) Unlisted packages are showing in results
When a version is unlisted the package should no longer show up as a
version. An issue was introduced by 43096df, which was meant to
simplify some of the Display Package page and pull versions to a
partial view, but the parts for keeping unlisted packages from
showing up was missed. This fixes that issue.