(GH-270) Do not allow HTML tags in markdown conversion
Before this change, a package creator could create a malicious script to
take advantage of the markdown conversion and run arbitrary scripts
with as little as a script tag stuffed in the description or
releaseNotes fields of the nuspec.

While the  possibility of something like this getting approved is low,
the attack vector might allow someone to gain access to another user's
cookies with a carefully crafted script and be sent to unsuspecting
folks by merely passing them a package page url.

To remove the possibility of arbitrary script execution or other
chances for HTML or script injection from packages, remove the ability
for the package contents to be interpreted as HTML tags and only
convert the contents to markdown after converting any "<" or ">" to
their html escaped alternatives.