(GH-115) Do not return rejected packages
If someone knows the exact version they are looking for, they could
still install a rejected package by passing the name *AND* the version
of the package. We've done the work in GH-211 (52db5134 and
bb86a6e5) to remove a copy/paste install box from google searches that
bring folks to unlisted/rejected packages. Now take it  one step
further and do not allow rejected packages to be installed.