(GH-268) Allow admins to exempt a package from testing
If a package id is not compatible with the testing for some reason,
it's best to go ahead and exempt the pacakge id from testing. This
means that the package testing in some respects will fall back to the
reviewer to do. We don't want to enable this very often.