(GH-275) Moderation email subjects more useful
The subject of the moderation emails has always been the same,
whether holding the package for further changes or when approving
the package with comments. The email subject should immediately give
the maintainer an idea if they need to take action or if the message
is simply an FYI.

- (GH-278) Approved with comments is only if there are review comments
and not just a status change comment.
- (GH-277) Allow moderator not to have the submit status change when
only commenting on a package under review and not holding the package.