(GH-302) Remove response for IDataServicePagingProvider
Due to 6afd749 a line was introduced to make a request determine if the
site implemented IDataServicePagingProvider return the implementation.
However FindPackagesById doesn't use a search filter so it was
prematurely turned on. It causes packages that would use a skipToken to
endlessly request the same url and appear to hang.