(GH-315) Cleanup Package API
Provide a way for an automated system to provide a reminder for
packages that are in waiting for maintainer status that have been in
that status more than 20 days. And after 35 days (15 days after the
noticed is sent) if no progress has been made, it should automatically
close (reject) the package with a note stating how folks can pick it
back up when they are ready.