(GH-322) Moderation should not show unlisted message
When viewing packages under moderation, the message should not show the
unlisted message. Instead, enhance the moderation warning message to
state that the package won't show up when searching for it and it is
not installable under regular means. Also enhance the moderation
warning a bit so it is clear that folks are on their own when
installing a version under moderation.