(GH-343) Cleanup rejection should assign reviewer
When the package cleanup is rejecting a package, it should assign the
review date and the reviewer. Otherwise the system makes it appear as if
the package was rejected on last reviewed date of however long ago that
was. Also having a quick way of knowing that the package was rejected
by a service automatically will make it easier to query/unreject later.