(GH-349) Rerunning tests makes pending status
If a moderator re-runs the automated checks on a package on which those
checks have failed previously, and the re-run indicates no problems -
i.e. all automated checks succeed - then the package should move to
Submitted (Ready) just as if the checks would have succeeded on the
first run.

Allowing the status to go back to pending will allow it to move into
another status automatically, even if that status is waiting for
maintainer to take corrective action.