(GH-376) Implement PowerShell Redirection Fix
PowerShell v2/3 caches the output stream. Then it throws errors due to
the FileStream not being what is expected. Fixes "The OS handle's
position is not what FileStream expected. Do not use a handle
simultaneously in one FileStream and in Win32 code or another
FileStream."

Running this in newer versions of PowerShell has been tested not to
cause any adverse affects, so it is presumed to be safe.

The fix comes from
http://www.leeholmes.com/blog/2008/07/30/workaround-the-os-handles-position-is-not-what-filestream-expected/
plus information in the comments that actually makes it work.