(GH-395) Allow html tags in review form
During the review process, folks will send parts of the nuspec, which
is xml, through asking questions. However the form errors every time
due to RequestValidation. This is counterintuitive for attempting to
contact folks and likely results in a lot of frustration from package
maintainers simply unaware that the site is failing due to the nuspec
tags they are putting in the page. The site should handle that better,
converting the output from the review message to encoded values.