(GH-386) Review messages clarify human contact
Make it clear that no human moderator sees a package until a maintainer
does something to fix the package or respond to the review comments.
Repeat this statement in a different way to catch folks who may not
understand what was meant by the first statement.

Until a maintainer leaves comments AFTER the automated review or both
pass, no human will see anything the maintainer may have stated prior
to those automated checks as the system will automatically roll through
the workflow process. If the checks fail, it goes back to the
maintainer to do something like respond to the review comments or fix
the package and resubmit.