(GH-406) Only show most recent nukg in scan results
Due to the many to many join, we don't have a great way of removing no
longer relevant results without possibly deleting it for other packages
that may use the same binary. Due to that sometimes on resubmission we
will see a package with multiple nupkgs in the results. We really only
want to see the latest one.