(doc) update how wiki generates code blocks
Use what highlight.js and the rest of the highlighters use to convert
code blocks to highlighted code. This means the language class is in
the code tag inside of the pre tag.

Remove highlighting for pre/code blocks that are sh or bash, as those
come out incorrectly, especially related to apostrophes in comments.