(GH-426) Use highlightjs instead of SyntaxHighlighter
Syntax highlighter has been great, but it has a limited number of
languages that it supports and it doesn't appear to be maintained at
the same level that highlight.js has been. There is a huge range of
support for languages, styles, and automatically determining languages
when it looks at a code block. Use it instead for the blog