(GH-424) Set 150MB Limit for Package Size
Package size was priorly 28.61MB as a default based on
requestFiltering's maxAllowedContentLength. This was probably not large
enough in some cases for Chocolatey packages. With a move towards the
ability to embed more software into packages when available, allow the
package size to be larger. When packages are over 150MB, provide better
guidance. If they go over the hard limit of 200MB, the request may come
back as a 404 still, but this will hopefully be seen less of the time
for folks.