(GH-509) Install.ps1 uses env vars to find install location
Sometimes Windows is not installed on the C drive. Previously the install script
did not account for this. Replaced reference to ProgramData with the
environmental variable for it in Install.ps1 with a fall back to
`$env:SYSTEMDRIVE\ProgramData\Chocolatey` if that method doesn't return the
right value. In most cases environment variables will return the correct value,
but there are a small number of cases where they will be overloaded. In those
cases, using a built-in environment variable like `SYSTEMDRIVE` will ensure that
at least the default Chocolatey location for nearly all instances will be found.