(GH-427) update version and pre switch
Using one dash with full name commands can produce weird warnings that look like
errors in newer versions of Chocolatey. Older versions of Chocolatey (pre 0.9.9)
required the parameters to be passed with a single dash (e.g. `-version`).

It's been two years since the release of 0.9.9, so folks should be on the newer
versions now.