(GH-512) Use non-BOM install resources
Not sure this is actually a PowerShell issue or just that the parameter
`-UseBasicParsing` causes any file that is downloaded to be treated as
ASCII, which in this case means it will be unable to work around the
BOM characters at the beginning of the file. `UTF-8` was changed to
`UTF-8 w/BOM` to account for
https://github.com/chocolatey/choco/issues/1225, but that seems to have
highlighted one additional issue - recognizing the type of script being
downloaded on `Invoke-WebRequest`.