(GH-514) Image 404 should not results in browser security warnings
The image caching and display service ensures that all images are
predownloaded so that there is no content being shown on the packages
pages except local content that is secured. However when that content
results in a 404 error, it was returning that original url and it
should return null instead.

Another fix dropped in here is to ensure when caching images is shut
off, it will use the original url instead of trying to build a path.