(GH-474) Separate package files from package results
Previously, package files were included with package information. This
was set to allow simpler query structure, but unfortunately it
increased load time and cache size for packages. With cache sizes, it
may be contributing to cache sizes too large for memcached. Separate
out the files and cache them per package version. This reduces the
cache sizes overall and will improve load time on pages for packages
that have lots of files.