(GH-522) Disable AutoFormsAuth / SimpleMembership
After reviewing some known issues in the release notes for ASP.NET MVC
(https://docs.microsoft.com/en-us/aspnet/whitepapers/mvc3-release-notes#0.1__Toc274034230),
it was determined that there is a known issue that causes Forms
Authentication to always redirect unauthenticated users to
~/Account/Login, ignoring what is set in web.config authentication
section. It is unknown why this hasn't hit the site before, but now it
is appearing. It is likely this could be causing other issues as well.

Ensure that autoFormsAuthentication is set to false and set
SimpleMembership false as well.