(GH-528) Add ac.uk + Others for Student Discount
More email addresses can be linked for student discount "self-service".
Those that end in ac.uk are typical for universities in the UK. Email
addresses for K-12 institutions in the US have k12.<stateabbrev>.us and
can be handled automatically as well (even though there may not be very
many of those).