(GH-548) make package results the default
The drop downs in the packages reload the page and then you have to
click back into the packages tab. That can get much more annoying in
practice, more so than the desire to want docs search results to show
first. Ensure packages are the selected tab on page load.