(build) Fix Website to Build with VS2017
Fix Website project to allow Visual Studio 2017 to build when using
Visual Studio 2017. For some reason the proper version number was not
being provided so it was looking for version 11 instead of 14. As most
folks would be using VS2017, check for existence of
Web.Application.targets path first, then if not found, use VS2017
MSBuild Web.Application.targets path instead.

Also handle cases with cassette.targets where spaces could be in the path for a
user so that builds still work properly.