(layout) Add CSS Libraries for Redesign
These libraries add styling to the new PackageLayout.cshtml, which will
be used for the redesign process.

- The bootstrap.css and style-base.css will be eventually included site
wide, while the packages.css will only be used in the packages area.

- As part of the redesign process, future gulp task will purge
unused css in the bootstrap css file, and then bundle together
with style-base.css. This will reduce the file size by at least half, if
not more.