(layout) Redesign of Package Details Page
The file "DisplayPackage.cshtml" is what gets displayed for each
individual package. The new design uses bootstrap components  and
allows for faster, easier use.

The files section has been updated to use the new Prism library for
syntax highlighting, and will only be applied when the file is clicked
on to be shown. There are now parameters in place that determine
when files will be expanded on load or hidden. When there are
too many files that are expanded at once, it greatly slows down the
performance of the website. In moderation view, there is now an option
to "expand all files" or "collapse all files", which may come in handy
when reviewing packages.

The comments section has been improved for easier readability and
updated to a more modern look and feel.

"custom.js" includes javascript that will be used site wide, while
"package-details.js" includes javascript only used on the
"DisplayPackage.cshtml" page.