(layout) Redesign of User/Login area
Files in the Views/Users folder now use the new "MinimalLayout.cshtml".
Design of this area has been updated to reflect the look of the new
Packages area.

-"MinimalLayout.cshtml" is a lighter weight version of
"PackagesLayout.cshmtl" that does not contain all header/footer elements
as they are not needed for this portion of the website.

- Account.css contains any styles that are specific to these pages.
Minor updates to style-base.css will also be used on these pages, but
may also be used elsewhere on the website.