(build) Add necessary Layout reference
These files remain unchanged except for the addition of "Layout =
"~/Views/Shared/OldLayout.cshtml" that is defined in the header. Right
now, this is needed because of how the website code is split into
separate themes. Once these pages have been converted to the new layout,
this line will no longer be necessary.