(layout) Redesign of Account overview page
Updates layout to using the new "PackagesLayout.cshtml" file. All
functionality remains the same, but now has a more modern look and
feel. It utilizes the new prism.js library for syntax highlighting and
Clipboard.js for copying.

-CSS styling for the "copy to clipboard" box has been moved from
packages.css to style-base.css as this snippet could be used throughout
the website.