(layout) Redesign of Edit Package Page
When the "Edit Package" icon is clicked from the "My Packages"
(Package.cshtml) page, a pop-up will now appear with info that was
previously on Edit.cshtml. It will give them the option to proceed with
uploading a new version of the package or canceling the process. If they
click the button to "Upload Updated Package" they will be redirected to
the upload package page. If they click cancel it will leave them on the
current page.

- This eliminates the need for the separate "Edit.cshtml" page.

- One line of CSS was added to style-base.css to make pop-ups appear
further down on the page.