(optimize) Extend optimization with gulpfile
Gulp is now in place working in conjunction with cassette for the time
being. Gulp will run before the build and then cassette takes over.

With the redesign of the packages area of chocolatey.org, new css
libraries were added. Gulp was implanted mainly to purge unused css
within these libraries, so that the size of the website stays small and
optimized. Gulp will also now compile Bootstrap from source within the
node_modules folder, and combine it with custom styles that were
previously located with style-base.css.

The style-base.css file has been converted into scss, and placed with
the scss folder. This allows custom variables to be set upon compilation
of Bootstrap, so that things like background colors and button colors
can be set initially within Bootstrap instead of overriding them. This
reduces the amount of custom css that is required and cleaner code.

gulpfile.js will first compile everything withint the scss folder and
create a new file, "chocolatey.slim.css". This file holds the main
chocoaltey theme and will be used on all pages of the redesign.

The gitignore file has been updated to include node_modules, and the
new chocolatey.slim.css which will always be generated on build.