(optimize) Update cshtml to use purged css
With the implementation of gulp, bootstrap.css is now compiled from scss
combined with custom styles that were previously in style-base.css and
now combined into one file, chocolatey.slim.css. This new file contains
only css that is being used on the website and not the entire Bootstrap
library. It will be included on all pages of the redesign of
chocolatey.org.