(layout) Redesign of Pricing page
Redesign features the new Chocolatey layout and theme utilizing the
bootstrap library. A new CSS file has been added for the "pages" area
of the website, and may be later included on other pages such as the
Documentation area or Blog.

Custom.js has been modified to include smooth scrolling on anchor
links, instead of it jumping abruptly to the designated area on the
page.

-Minor css adjustments to account.css, mostly adding
specificity to selectors or to accommodate styles that may be used
throughout the entire website.