(layout) Redesign of Blog List View
All blog posts have been edited to add in an image that will display in
the blog list view (/blog). Going forward, each blog should have an
image listed in the metadata, and there should be no inline styles
associated with it.

If the blog post does not have a desired image, javascript will now
insert the a default img of the Chocolatey Icon Logo.