(layout) Update of redesign of Packages Overview
Individual packages on the packages overview page have been converted
over to use bootstrap "cards". Bootstrap "cards" are used throughout
the entire redesign, so these were updated to stay consistent with the
rest of the theme. A "Learn More" button has been added that will take
the user to the individual package.