(GH-645) Add ContactAdmins view model
The Report Abuse form and the Contact Admins form previously shared the
ReportAbuse view model. With the addition of the required check box to
the ReportAbuse view model in a previous commit, this made the Contact
Admins form unable to submit. The Contact Admins form was expecting a
required check box element, when there was none.

A new view model has been created specifically for the Contact Admins
form, and all appropriate files have been updated to reflect the change.