(optimize) Template navigation for easier editing
The top and bottom navigation elements have been templated out so that
if the items need edited, now there is only 1 file to edit as opposed
to several. This cuts down on user error of making an update to one
file, but not another. The top and bottom alert message are now in
their own file as well, so that they will also only need updated in one
spot in the future.