(layout) Top Navigation Redesign
The top navigation has been updated to reflect the redesign of
Chocolatey.org. Now, when clicking on a navigation link, a dropdown
sub navigation will appear with more links in the category. Minor class
changes were made to pages that contain a class "header" to accomidate
styling of the new top navigation.