(layout) Content Redesign of Home Page
The content of the home page has been reworked, which requried a few
additional design updates. Logos have been added of companies who are
partners or integrate with Chocolatey, as well as a testimonial
carousel. The header has also been modified with a gradient background
image. This look will be replicated across multiple pages.