(maint) Rename "Kickstarter" to "Businesses"
Simply renames the "Kickstarter" images folder to "Businesses". This
was done because logos within this folder will potentially be used
throughout the website, which then doesn't make sense for them to be
located under a "Kickstarter" folder. In result of changing this folder
name, the image paths on the Kickstarter page have been updated.