(layout) Addition of Resource Library
This area of the website has been added to house resources that may be
useful to any type of Chocolatey user. It contains Case Study videos,
Customer Success Stories, and other demo videos, along with a "home"
page for getting to Documentation, Blogs, and the "Getting Started"
course.

The README.md file explains the Tags and Types associated with each
resource which will need to be updated as this area grows.

A new resource is added by including a new .md file within the "files"
folder and assigning appropriate values to each field. An exisiting .md
file can be used for a good starting point for any resource. Once a new
.md file is added and approriate Tags and Types are assigned, it will
automatically populate into it's designated area.