(layout) Addition of Community Page
The Community page has been added for a "home" of how users can get
involved with the Chocolatey community. It provides links on Chocolatey
discussions, how to contribute, and where to follow us online.

Minor css adjustments to _base.scss and _icons.scss were made to
support the Community page.