(layout) Addition of Products Page
With the addition of the new Products page, the old Business page has
deleted. In result, links throughout the website have been updated to
point to the new Products page wherever it was appropriate. The Products
page gives users a general overview of each Chocolatey edition offered,
and may contain links to other pages for more information.