(layout) Addition of Compare Page
A new page has been added specifically for the Compare table that was
previously located on the Pricing page. In result of moving this table
to it's own page, it has been deleted from the Pricing page.

Links throughout the website have been updated to the new route url to
point to the new page.