(layout) Redesign and Addition of Discount Form
The old Discount Form has been updated to reflect the current style of
the other contact forms.

A new Discount Form has also been added with the redesign of the Contact
area. Instead of using one general form and having to select "Student
Discount" from a dropdown, users will now be redirected to this new form
specifically for discounts.