(layout) Make tabs responsive on mobile
The following changes allow tabs to become buttons on mobile devices.
This is more user friendly, and ensures the user is able to access the
information they need. All tab css has been moved to a new scss file
"tabs.scss" and compiled in the main chocolatey.slim.css file which is
included thoughout the chocolatey.org. Minor changes were made to
.cshtml files to allow the tabs to turn into buttons on mobile.