(maint) Delete unneeded Layout references
During the redesign process, the layout was defined on each page due to
the website being split, and using multiple different layouts depending
on if they had been redesigned yet or not. Now that the redesign is
complete, the main "_ViewStart" file has been updated to define the
appropriate layout being used, which means each individual reference
can now be deleted.