(maint) Delete unneeded references to closeable.js
The javascript file closeable.js is only needed on pages that users can
access packages. The reference to this file however, was on almost every
page. The reference has been removed from pages that do not need it,
and appropriate code has been placed on files that do. This reduces
unnecessary server request and cleaner markup.