(layout) Add Status Box Details and Update Icons
A new "status box" has been added to DisplayPackage.cshmtl, so that
when a user goes to an individual package, they will now easily be able
to see the status of verification and validation testing.

The old ball icons have been deleted, and replaced with a CSS variation.
The new css status ball is split in the middle, so that half of it
reflects the status of verification testing, and the other half reflects
the status of validation testing. In this way, the ball could be 2
different colors. Also, in the event that more test are added in the
future, this css ball can be modified to be split into however many
sections needed, to represent the total number of tests and statuses of
each.