(layout) Optimize code used for Search results
Previoulsy, when someone did a search on Chocolatey.org, depending on
what page they did the search on, it would use different controllers
and display items on a different .cshtml file, although the code/page
was exactly the same. Now, no matter where the search is done from, it
is using the same code and loading results into the same .cshmtl file.