(GH-619) Fix "Previous, Next" Pager Links
The "Previous, Next" links will now either go to "/search" or
"/packages" based on what URL the user is already on. Logic has been
added in to detect the current URL. The layouts are also now using the
same .cshtml file and the look will remain consistent no matter what URL
the user is on.