(GH-228) Allow Moderator Queue to be sticky
If a user has a package role, they will now be able to set the
Moderator Queue to their default view. While in the Moderator Queue,
and the user is logged in, they will see a toggle switch on the top
right corner next to the search filters. This sets their default view,
and will remain set until the user clears their local storage.