(layout) Adjust styling of package communication
The package communication area is converted from markdown, and then
styled with javascript to make it easier to read. The previous styling
was wrapping each comment in a "li" element, which was causing problems
if there was another "li" element inside the actual comment. This
changes that element to a "span" element, which will no longer interfere
with the individual comments.

Minor additional addjustments made to package-details.js for
optimization.