(layout) Update package comments area editor
A new library named simplemde has been added to the DisplayPackage page.
This adds a wizzywig type editor the comments section. It allows users
that may not be too familiar with writing in markup to do so. It
includes a toolbar that users can easily select what markdown component
they are wanting, and also a live preview section so users can preview
what their comment will look like before submitting it.