(layout) Add written testimonials to resources
A section for written testimonials has been added to the resource home
page. This section may transform as more testimonials are added
in the future. A snippet of javascript has been added which will
shuffle the order of the written testimonials on page load. This
feature will come in handy once more testimonials are added, that way
users get a chance to read all of them if they happen to view the page
more than once.