(build) Convert Documentation to use Markdig
The documentation area is now using Markdig to convert .md files as
opposed to MarkdownSharp. After the files are coppied stright from the
choco-wiki submodule, they are placed in a folder inside the
Documentation area. The files are converted on the fly by the
DocumentationController instead of being converted to cshtml.