(layout) Update cshtml to accommodate viewports
With the update of the viewports back to Bootstrap standards, some
cshtml needed to be updated. While it looks like a lot of updates, the
main thing that has been done in these files is unnesting them from
containers that would in tern make their appearance a smaller width.
With the reduction of global containers, these are no longer needed. Not
adjusting these resultes in content that is too "squished".