(layout) Mobile video pop-up
Normally, when the page is scrolled on a video pop-up, it will dismiss
the pop-up and stop the video. This works great on desktop but not so
much on mobile devices. Logic has been added to make this feature only
available on desktop. While on mobile, the user must manually close the
pop-up.