(GH-698) Search when default to moderation queue
A simple change to the javascript has been made to allow searching the
normal package view, even if the default view is set to the moderation
queue. Now, the javascript looks to see if a search is being performed,
and if so will not redirect the user back to the moderation queue. It
will perform the search as intended.