(layout) Enable package list view/preferences
A new button on the package list view has been added that will bring up
a modal presenting options to switch the package grid view to a list
view. For now this option is cookie based. The other option to make the
moderator queue the default view has been moved into this model to keep
user preferences all in one area.

The "ListPackages" files have been modified to accomidate this new view
and reuse code on both the grid and list view.

A line in the PackagesController has been added to support making the
modorator queue the default view if the user has this option enabled.