(layout) Improve testing results UI
The order of the test displayed by the ball icon next to the package
title have been switched around. Since Validation testing is ran first,
it is represented on the left, and will show up first in the drop down
box on each package page. The Verification testing is represented on
the right side of the ball, and under Validation testing on each
package page.

In addition to moving the order of the test, additional logic has been
added in to give more specific testing status if one check in exempted
from Verification.