(layout) Package warnings are available on mobile
Previously warnings about  if a package was a prerelease or not yet
uploaded to CDN were included on desktop but not on mobile. The correct
text has been added to show these warnings on mobile now. The font size
for desktop has also been brought down and white space removed.