(layout) Make collapsed elements navigatable
A snippet of of javascript has been added so that when a user clicks a
link that navigates to a collapsible element, it will open the
collapsed element and scroll to it. Previously if the collapsed element
was not already open, it would scroll to the top of the page.

This snippet was previously only on the pricing page, and has been moved
to the custom.js script to be used site wide.