(layout) Enable tab navigation by URL
A javascript snippet has been added that will open up closed tabs based
on the current URL. Also, it will update the URL when clicking on a tab,
that way the URL can be coppied and given to someone else so that they
can navigate to the correct tab.

Added the bouncing ball loading screen to the install page to prevent
the user from seeing the javascript opening tabs on the load.