(layout) Change "View" to "Show" on package page
Updates the word "View" to "Show" on mobile devices for the links
dropdown container. This was updated to stay inline with the verbiage
of the rest of the website. On click, "Show" will turn to "Hide". When
clicked again "Hide" will be changed to "Show". This is taking use of
javascript already being used on other parts of the website.