(layout) Global search in top navigation
A search bar has been added to the top navigation. If the user is on
the home page, this search box will search through packages and
documentation and present the user with both results.

If the user is in the documentation area and on a mobile device, this
search box will only search through the documentation and keep the user
in this area.